+++ CHANGELOG.md
@@ -8,7 +8,12 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ### Changed
 
+* Upgraded Nextflow plugin version to 1.6.0
 * Nextflow - updated default instance types based on destination region
+
+### Fixed
+
+* Nextflow errorStrategy retry ends in 'failed' state if last retry fails
 
 ## [339.0] - beta